requirements.txt
fastapi[standard]
python-dotenv
sqlalchemy
psycopg2
passlib
bcrypt==4.0.1